+++ src/components/curation/Mechanism.vue
@@ -1,6 +1,6 @@
 <script>
 import MechanismEvidence from "./MechanismEvidence.vue";
-import CustomToolTip from "../tooltip/CustomToolTip.vue";
+import ToolTip from "../tooltip/ToolTip.vue";
 import {
   MechanismAttribs,
   MechanismSynopsisAttribs,
@@ -25,7 +25,7 @@ export default {
   ],
   components: { 
     MechanismEvidence,
-    CustomToolTip
+    ToolTip
   },
   data() {
     return {
@@ -57,7 +57,7 @@ export default {
             <div class="row g-3 px-3">
               <p class="mb-0 fw-bold">
                 Badoyni et al probabilities
-                <CustomToolTip
+                <ToolTip
                   toolTipText="These scores were calculated using a tripartite statistical model based on diverse gene and protein level features 
                   that are known to discriminate between molecular mechanisms. It is made up of machine classifiers trained to predict 
                   whether human protein coding genes are likely to be associated with dominant-negative, gain-of-function, or loss-of-function 
@@ -73,7 +73,7 @@ export default {
                     <tr>
                       <td width="60%">
                         Gain of Function (pGOF)
-                        <CustomToolTip
+                        <ToolTip
                           toolTipText="Badonyi and Marsh pGOF:
                           the probability that the protein is associated with a gain-of-function disease mechanism 
                           (pGOF). Gain-of-function mechanisms are characterized by an altered or newly appeared function in the mutant protein. 
@@ -90,7 +90,7 @@ export default {
                     <tr>
                       <td width="60%">
                         Loss of Function (pLOF)
-                        <CustomToolTip
+                        <ToolTip
                           toolTipText="Badonyi and Marsh pLOF:
                           the probability that the protein is associated with a loss-of-function disease mechanism (pLOF).
                           A loss-of-function mechanism is characterised by complete ablation of the function of the proteinGenes
@@ -107,7 +107,7 @@ export default {
                     <tr>
                       <td width="60%">
                         Dominant Negative (pDN)
-                        <CustomToolTip
+                        <ToolTip
                           toolTipText="Badonyi and Marsh pDN: 
                           the probability that the protein is associated with a dominant-negative disease mechanism (pDN). 
                           Dominant-negative mechanisms are characterized by the mutant protein directly or indirectly disrupting the function of the wild type protein.
